# Improving sort performance by handling AbstractVector and rev=true by counting sort