Currently, sorting integers with small range in a `SubArray` is slower than the same case with a `Vector`:

```julia
julia> xs = rand(0:9, 1000);

julia> @btime sort!(xs) setup=(xs = $(copy(xs)); copyto!(xs, $xs));
  1.559 μs (1 allocation: 160 bytes)

julia> @btime sort!(xs) setup=(xs = view($(copy(xs)), :); copyto!(xs, $xs));
  5.136 μs (0 allocations: 0 bytes)
```

First commit "fixes" it:

```julia
julia> @btime sort!(xs) setup=(xs = view($(copy(xs)), :); copyto!(xs, $xs));
  2.745 μs (1 allocation: 160 bytes)
```

It's not as fast as the `Vector` case but I think it's mostly due to non-optimized `extrema`:

```julia
julia> @btime extrema(xs);
  194.520 ns (1 allocation: 32 bytes)

julia> @btime extrema(view(xs, :));
  988.615 ns (2 allocations: 80 bytes)
```

In the second commit, I tweaked the counting sort a bit to handle `rev = true` case:

Before:

```julia
julia> @btime sort!(xs; rev=true) setup=(xs = $(copy(xs)); copyto!(xs, $xs));
  5.754 μs (0 allocations: 0 bytes)
```

After:

```julia
julia> @btime sort!(xs; rev=true) setup=(xs = $(copy(xs)); copyto!(xs, $xs));
  1.670 μs (1 allocation: 160 bytes)
```